requirements.txt
# Flask e extensões
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1

# Banco de dados
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Segurança e rate limiting
Flask-Limiter==3.5.0
python-dotenv==1.0.0

# Processamento de dados
pandas==2.1.4
openpyxl==3.1.2

# Markdown
markdown2==2.4.10

# Servidor de produção
gunicorn==21.2.0

# Utilitários
python-dateutil==2.8.2
Pillow==10.1.0
pytz==2023.3

# Desenvolvimento (opcional)
# matplotlib==3.8.2
# seaborn==0.13.0
